odd that i had to pull before push? interesting
